requirements.txt
flask
PyJWT
python-dotenv
jsonify
Flask-Argon2
flask-jwt-extended
flask-jwt-extended[asymmetric_crypto]
Flask-SQLAlchemy
psycopg2-binary
email-validator
redis
flask-cors
pytest
flask-swag